feat: 회사 등록 테스트코드
